Improve how we deal with static resource.
We now have .gz versions of the resources, which we send to clients when we
can. We also calculate digests of the resources based on a hash of their
contents and use this (currently) for ETags and (in the future) for other
cache control headers.

The digest stuff was a hell of a pain to set up but it's actually pretty cool
now; plus it's a little fancy which is always fun. One interesting question:
is the Closure optimisation of the JS that happens idempotent, or is the
content-based hashing effectively equivalent to my traditional Git SHA-based
hashing?